Creature class: Implementing move method + declaration and definition of helper method to find the target
+ silence implicit casting warnings

+++ Evol'artist/src/main.cpp
@@ -5,6 +5,8 @@
 //  Created by Emma Barme on 28/09/2017.
 //  Copyright GNU General Public License v3.0
 //
+//  Last modified by Emma Barme on 29/09/2017
+
 
 //main function for tests
 
@@ -16,7 +18,7 @@ using namespace std;
 
 int main(){
     
-    srand (time(NULL));
+    srand ((unsigned int)time(NULL));
     
     cout << "Hello World!" << endl;
